Add a --dry-run option

+++ lib/automedia/move.ex
@@ -5,17 +5,18 @@ defmodule Automedia.Move do
 
   require Automedia.Movable
 
-  def move(%Automedia.Movable{source: source, destination: destination}) do
+  def move(%Automedia.Movable{source: source, destination: destination}, options) do
+    dry_run = Keyword.get(options, :dry_run, false)
     path = Path.dirname(destination)
     if !File.dir?(path) do
       Logger.info "Creating directory '#{path}'"
-      File.mkdir_p!(path)
+      if !dry_run, do: File.mkdir_p!(path)
     end
     if File.regular?(destination) do
       Logger.info "Skipping move of '#{source}' to '#{destination}' - destination file already exists"
     else
       Logger.info "Moving '#{source}' to '#{destination}'"
-      File.rename!(source, destination)
+      if !dry_run, do: File.rename!(source, destination)
     end
   end
 end
